modifying Scoped Functions to return nullable and adding test cases
